fileingestor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
Pillow==10.1.0
imagehash==4.3.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic
pypdf==4.0.1